Create model_analysis
A place to add all the utility for the analysis of the model. (I would like to contribute to the project, so could I do this section? Your work so far is prolific and incredible, and I appreciate it, but I also would like to contribute. I'll avoid overwriting any of your code.)